move files (#39)
https://github.com/frolovdev/anyspec/issues/23

closes #39
