+++ src/language/index.ts
@@ -1,2 +1,7 @@
 export * from './predicates';
 export * from './ast';
+export * from './lexer';
+export * from './parser';
+export * from './source';
+export * from './token';
+export * from './location';
