+++ src/parserModel.test.ts
@@ -1,8 +1,13 @@
-import { ASTNode, ASTNodeKind, EnumTypeDefinitionNode, ModelTypeDefinitionNode } from './language';
+import {
+  ASTNode,
+  ASTNodeKind,
+  EnumTypeDefinitionNode,
+  ModelTypeDefinitionNode,
+  parse as defaultParse,
+  Source,
+} from './language';
 import { AnySpecError } from './error/AnySpecError';
-import { parse as defaultParse } from './parser';
 import { toJSONDeep, log } from './utils';
-import { Source } from './source';
 
 const parse = (source: string | Source) => defaultParse(source, { noLocation: true });
